desktop_backend/requirements.txt
langchain==0.2.5
langgraph==0.0.69
langchain-openai==0.1.14
requests==2.32.3
gitpython==3.1.43
python-dotenv==1.0.1
psutil==5.9.8
